improve table view (especially on mobile) with scroll-snap css
Scroll-away headers can be nice for increasing available
viewport space (especially on mobile) for content, but they
can also be quite annoying when they reappear before you want them to.

Adding some scroll snap CSS fu on table-view prevents the
header from reappearing when the user scrolls up to the top
to see the first row of table data.

With this commit, the scroll action has to be a bit more emphatic
to make the header reappear.  That's the intent anyway.

May need additional horse-whispering.