make scrolling table view less 'loosey-goosey'
It feels too easy to trigger the scroll-away header to reappear
in table view.

I had been using:

  scroll-snap-type: proximity

for the table container

But in this case, I think it's okay to use:

  scroll-snap-type: mandatory

and align individual table rows to the start of the snap
boundary.

Unlike photo-view where it is entirely possible for the user
to want to view a partially visible city-card, especially
in landscape on mobile, the individual rows in table view take
up little vertical space relative to the available viewport.
So 'mandatory' is tolerable and will actually prevent the
scroll-away header from reappearing if the user ballistic-scrolls
to the top of the table.