table view on desktop and mobile still triggers scroll-in header too easily
Adding some padding to the table scroll container seems to give the behavior
I desire ... which is to allow a ballistic scroll back to the top to
see the first row of data, but /not/ trigger the header to reappear.

We only want the header to reappear if they then drag the mouse down
from that position.  Adding some vertical padding seems to get us out of
the jitter zone (at least on desktop).

We'll see on mobile. :-)