make chart-view scrollable now that bottom appbar is fixed and behaving
While I was battling the 100vh webkit misery

(see: https://allthingssmitty.com/2020/05/11/css-fix-for-100vh-in-mobile-webkit/)

I had disabled scrollability of chartview entirely.  But my proclivity
for vh units meant that city labels were invariably getting clipped
(especially on mobile) because vh doesn't factor in the browser's bottom
appbar.

With Matt Smith's fix for that working, it's safe to turn on scrolling
again.  I'm also repositioning the leged below the graph so that in the event
of clipping, it's that and not the more interesting city lables.

Chart view is gonna get reworked some day soon so this is really
just tactical, shortterm goodness.