Revert "nix the headers on chart view; use floating menu instead"
This reverts commit da02cc6dff4a7142ac01728ce22aa424e8315cb4.

I'm reverting this fix because it introduces a regression with map-view tooltip not
appearing when internet is down and chart-view is selected.

The intent was to communicate, through explanatory tooltip text, /why/ the map-view icon is grayed out
when the internet is unavailable ... basically, map-view has a hard requirement on internet because
it does a lot of ajaxy fetches to render the map.

I zero'd in on this with git bisect.  It allowed me to find the sketchy commit out of a haystack
of 73 commits in only 7 steps. :-)