make network availability event-driven with window.online/window.offline listeners
Map-view requires internet to function.  Prior to this commit, I perform a synchronous
check at start-up and gray-out the map-view button on the results page if
internet is down.  I also provide some tooltip language for user to reload when
internet is back.

That's better than nothing, but still a bit lame.  The app should just /detect/
if/when the network comes back.

With this commit, the results page adds window event listeners for "online" and "offline"
status.

It's true these events really just respond to the link-layer (as opposed to true transport-layer connectivity).  But paired with the synchronous check at start-up, offer a pretty decent statusfor network availability.